fixing report value

+++ src/main/java/JEstebanC/FastFoodApp/service/ReportServiceImp.java
@@ -217,14 +217,14 @@ public class ReportServiceImp implements IReportService {
 
     private ReportSalesMonthlyDTO convertReportSalesMonthlyToDTO(Map<String, Object> salesMonthly) {
         ReportSalesMonthlyDTO reportSalesMonthly = new ReportSalesMonthlyDTO();
-        reportSalesMonthly.setMonth((Double) salesMonthly.get("month"));
+        reportSalesMonthly.setMonth((long) ((BigDecimal) salesMonthly.get("month")).intValue());
         reportSalesMonthly.setTotal(((BigInteger) salesMonthly.get("total")));
         return reportSalesMonthly;
     }
 
     private ReportSalesWeeklyDTO convertReportSalesWeeklyToDTO(Map<String, Object> salesWeekly) {
         ReportSalesWeeklyDTO reportSalesWeekly = new ReportSalesWeeklyDTO();
-        reportSalesWeekly.setWeekday((Double) salesWeekly.get("weekday"));
+        reportSalesWeekly.setWeekday((long) ((BigDecimal) salesWeekly.get("weekday")).intValue());
         reportSalesWeekly.setTotal(((BigInteger) salesWeekly.get("total")));
         return reportSalesWeekly;
     }
